requirements.txt
Django
djangorestframework
django_rest_apikey
django-rest-framework-nested
django-rest-swagger
MySQL-python
pycrypto
